avoid personal commentary in the spec itself
Nico wrote this note about message composition that is a Reply to
encrypted messages:

   Nico: IMO it has to. IMO it is key that a reply to an encrypted email
   by default always is encrypted.
   Especially because otherwise one state change of multiple recipients
   will disable encryption for all.
   The question is whether this is part of the autocrypt recommendation
   to the MUA or whether this is a recommended behavior for MUAs for how to
   deal with the autocrypt recommendation.

This sort of commentary belongs on-list, in git notes, in pull
requests, or in git history, but probably not in the text that is
presented publicly.  I'll push a branch that tries to incorporate a
concrete suggestion to resolve this next.