propose Autocrypt recommendations for replying to encrypted messages
Replying to encrypted messages in the clear is considered a faux pas
in the traditional e-mail encryption world.  As an opportunistic
scheme, Autocrypt might not always be able to avoid this.  However,
Autocrypt should avoid doing this if it knows the keys for all
recipients targeted by the reply.

There are some sticky corner cases that the user might encounter when
replying to an encrypted mail if they do not already have the
necessary keys, unfortunately; handling those cases specially is
difficult, especially if we want this to be convenient and
comprehensible for most users.

I welcome alternate proposals for how MUAs should deal with this
situation!