only discourage sending encrypted after a timeout
If an autocrypt mail was seen within the last month we allow encrypting
without a warning.

We are trying to destinguish two cases here:
 * the recipient is using an autocrypt and a naive client in parallel
 * the recipient stopped using autocrypt

In the former case we expect to see autocrypt headers from time to time.
We don't want to switch between different behaviours frequently.

In the latter case the recipient may still see unreadable for a
month but we will discourage sending encrypted mail afterwards.