further improve MUSTs/SHOULDs
@azul and @hpk42 pointed out that we had MUST language about being
capable of generating RSA 2048, but not about actually using them for
this account.

The MAY around 25519 seemed a little too strong -- effectively
encouraging people to consider it.

This update (made by @dkg and @azul) should strengthen the MUST use,
and changed the MAY to acknowledge the use case without hinting at
25519.