cleanup: remove encrypt button todo
We already say that the client MAY hide the button.
This is not a SHOULD - maybe it was in the past.
Either way i think we are pretty clear that it's up
to the client to decide wether to display the button
when disabled. We only say it SHOULD display the button
if 'available' or 'encrypt'.

fixes #105

in line with what @r10s comment.