Treat aliases as distinct accounts
Way back in #25, @hpk42 and @azul had a discussion about treating
aliases differently than separate accounts.

As our work converges on the idea that a single e-mail address is the
primary object label in Autocrypt, it seems to be easier to treat an
alias as a special case of a controlled account.  This effectively
inverts the guidance we have been giving in the past about Aliases.

The Autocrypt Setup Message should continue to work fine for different
keys for different aliases -- each Setup Message will have the From:
and To: fields marked with the specific alias that it refers to.

This change doesn't contemplate a situation where the MUA is
dynamically spawning new aliases on the fly all the time, but such an
implementation will interact poorly with Autocrypt anyway.

I've discussed this with both @Valodim and @azul and no one has
proposed a downside to this simplifying change to the spec, so i
thought i'd try proposing it.  Please feel free to reject the change
and explain why if you think it is likely to cause problems.

Closes: #178