tighten up authentication for travis push
We can configure the user home directory to know what server host keys
to expect (so that we don't just use our id_rsa to talk to arbitrary
servers).

We also unpack the id_rsa so that it should be findable directly by
ssh, which means our invocation of rsync can just be the default.