Re-frame introduction to the Autocrypt Recommendation
This is meant to be considered as an alternative to PR #224.

As i wrote in the discussion on that thread:

  IIRC, the intent of the initial paragraph was to permit clients to
  claim "Autocrypt" compatibility, while still experimenting with (A)
  new data sources or statistical metrics, and (B) new recommendation
  algorithms while still trying to minimize/avoid the unnecessary pain
  points typically inflicted on users by encrypted mail.

Hopefully this captures that intent better.

With the increased text here, i thought it was better to break out the
technical description of the structure of the Recommendation algorithm
(its inputs and outputs) into its own subsection, so that the
descriptive/normative/motivational text is distinct from the technical
specifics.