Refer to GnuPG bug report instead of tracking it in Autocrypt
We've identified a problem with GnuPG importing
symmetrically-encrypted data safely based on the passphrase.

This should be reported on the GnuPG bug-tracker, not held in the
Autocrypt repo. I've pushed as much detail as i think is relevant on
this problem to the GnuPG bug-tracker itself, so i think we can just
point there.

Closes: #239