another try to fix #248 but hopefully meeting more agreement :)
So this PR aims to cause more readable mails for Carol compared to
the current spec state (which, btw, references "gossip_key_timestamp"
which does not exist).  It should be at least as good as the current
spec state but gives parties receiving a gossip key a single
bit of information: whether the gossiping party got the key
directly or indirectly within the last ~35 days.

With this PR and the #248 scenario:

- Alice sends a Gossip-Header for Carol without the "fresh=true"
  setting in her group message to Bob & Carol

- Bob sends a mail using his own Carol key - Carol can read
  it because Bob has the correct key in the #248 scenario.

If instead Alice has a "fresher" key:

- Alice sends a Gossip-Header for Carol wit the "fresh=true"
  setting in her group message to Bob & Carol

- Bob will take Alice's key for Carol if he didn't
  receive a message from Carol in the 35 days before
  gossip_timestamp.

And a note regarding "contact privacy": the gossiping party leaks
whether she has gotten a mail from a peer within the last 35 days (by
sending fresh=true) along the mail.  However, without this PR anyone
who does not use the gossip key in their replies, leaks that
information as well.