Indicate that the MUA might also have access to other sensitive info
We're talking about protecting the secret key with a passphrase, but
we haven't mentioned that other secret information might be just as
relevant.

I want the MUA developer to remember that other secret data they hold
(e.g., the contents of encrypted mail) could be just as risky to leak.