avoid loaded/imprecise "full-disk encryption" term
What some people call "full-disk encryption" is often not actually
over the full disk, and it doesn't provide as much fine-grained
control as many users want/need.

For example, if you want to be able to use your computer for a
non-mail purpose without exposing your inbox to any process that
happens to be running, actual full-disk encryption is too
coarse-grained of a control.

In addition, disk encryption has nasty failure modes if (once
decrypted) the underlying files have extremely loose permissions such
that any process, under any user account can read the files.  So it
must be combined with other forms of access control to be safe.

So while FDE is a baseline of something that all computers should
have, it's also a mistake to imagine that it solves all the problems
of sensibly-scoped access control.

This updates the text to acknowledge that thinking about access
controls around the filesystem level (and in other places) are
necessarily concerns that might be relevant for a security-sensitve
mail user agent.

We also echo the "and other sensitive data" refrain from the earlier
paragraph as a reminder that the keys may not be the only thing worth
protecting.