avoid inappropriate MUST in secret key storage section
Even though MUST is correct, introducing it an a justificatory clause
seems unfair to the reader.  It comes across as looking like the
section on secret key protection is introducing a requirement relevant
to e-mail generation.

Instead, we should link to the requirement we're talking about and
avoid the use of normative language around behavior not directly
addressed by this section.