update optional state to current level 1
I removed the OPTIONAL: inside the message processing because
the whole process is optional - but once you do it none of the
steps are optional anymore. If you leave them out the count will
be wrong.

I also removed references to ``autocrypt_peer_state[A]`` rather
than updating them to ``peers[P]`` to improve readability and
make this less coupled to the higher level data structures.

We do not have a reset state anymore - therefore we now compare
``counting_since`` to the ``autocrypt_timestamp``. This is
effectively a detection of resets because we update
``counting_since`` on resets.