delete ui directory
This was a prototype we used during IFF 2016. It is archived here aswell:
  https://github.com/autocrypt/ui-autocrypt/releases/tag/v1.1-fixes

It served to discuss the configuration UI and illustrate how little ui
elements we expect autocrypt to have.

We are close to having actual implementations we can demo instead that have
realised autocrypt with a minimal UI.

In addition the prototype is not in sync with our current discussion and
might be confusing people.