Document disabling and destroying secret key material
We currently have no discussion about the details of disabling
Autocrypt, or of how to expose the user to the choice of destroying
secret key material.

This change introduces this necessary text, and also observes that it
is the simplest/shortest path to synchronizing secret keys.

This changeset lays the groundwork for option (c) as discussed in
https://github.com/autocrypt/autocrypt/issues/277.  A future changeset
can use this approach to deal with colliding Autocrypt level 1 MUAs in
a minimalist way.