website: put the actual mission first
When making a mission statement, the first sentence should be a simple
and clear statement of our goals.  It should not start with an
explanation of why we haven't reached them yet, or of the historical
background.  Those things are fine to include, but the first sentence
after "Our Mission" should be the mission of the project.