incremental improvements on @valodim's hotfix last night, aiming for:
- brevity
- trying to be legible and useful for both end/expert users and
  security-minded hackers

we've had two public waves of interest (first heise, hackernews now) and i
guess we will keep changing/adapting things as tides role in,
and in particular when more mail apps with l1 support are released ...
I think "single-click opt-in encryption" is a useful catchphrase
(the fact that we also care for group encryption is missing but
maybe someone can try that in a different PR?).  Moreover,
mentioning encrypted group communication and multi-device
support needs to be mentioned, as these are pain
points known to both existing users and developers.

I btw welcome the evolving practise of relatively quick merges --
if we evolve a practise that subsequent PRs can change things
equally easily :)