Angle dev-status table headers
This appears to work in more browsers, and is more readable than the
previous attempt.

Thanks to https://css-tricks.com/rotated-table-column-headers/ for
some hints and ideas.

I needed to wrap the column headers in *s to get an additional em
element in there to do the rotation as intended.

Future possible improvements:

 * lines between the headers that line up with the lines of the
   columns

 * make left-most column ("MUA/project") position: sticky

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>