Accomodate different versions of sphinx
Sphinx 1.8.5 generates table headers with italics as:

    <th><em>foo</em></th>

But Sphinx 2.2.1 apparently adds an additional <p> wrapper between the
<th> and the generated <em>:

    <th><p><em>foo</em></th>

This adjustment in the css selectors is trying to make things work
with both versions of generated output.

If we could depend on the 2.2.1 generation, then we could apply what's
currently on the <em> to the <p>, and use the <em> as the equivalent
of a span and try to get the line separators to work there.  But i
given that 1.8.5 is in debian unstable, i don't want to depend on
2.2.1 just yet.

It looks like sphinx 2.2.1 might also apply "vertical-align: middle"
to <th> elements, so we force that to "bottom" as well.